requirements.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
konlpy==0.6.0
wordcloud==1.9.2
matplotlib==3.7.2
sumy==0.11.0
nltk==3.8.1
pandas==2.0.3
numpy==1.24.3
lxml==4.9.3
# PDF 리포트 생성
fpdf2==2.8.3
# 이메일 발송
email-validator==2.2.0
python-dotenv==1.0.0